Cambios en los tipo de variable para adaptarlo al modelo

+++ src/main/java/ufro/ignaciou/taller1_backend/models/Counter.java
@@ -20,7 +20,7 @@ public class Counter {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_counter")
-    private Long idCounter;
+    private Integer idCounter;
 
     @Column(name = "nombre")
     private String nombre;
